feat: improved config and env override handling
- The following config parameters now support size suffixes:
  - query-initial-memory-bytes
  - query-memory-bytes
  - query-max-memory-bytes
- The following parameters are now range checked properly to prevent
  compaction issues and log corruption with values that overflow
  internal size limits:
  - aggressive-points-per-block
  - compact-throughput
  - compact-throughput-burst
  - max-index-log-file-size
- Fix potential loss of accuracy when displaying default parameter
  values when displaying help
- Fix potential panics for default parameter values when displaying help

+++ cmd/influxd/launcher/launcher.go
@@ -434,9 +434,9 @@ func (m *Launcher) run(ctx context.Context, opts *InfluxdOpts) (err error) {
 
 	m.queryController, err = control.New(control.Config{
 		ConcurrencyQuota:                opts.ConcurrencyQuota,
-		InitialMemoryBytesQuotaPerQuery: opts.InitialMemoryBytesQuotaPerQuery,
-		MemoryBytesQuotaPerQuery:        opts.MemoryBytesQuotaPerQuery,
-		MaxMemoryBytes:                  opts.MaxMemoryBytes,
+		InitialMemoryBytesQuotaPerQuery: int64(opts.InitialMemoryBytesQuotaPerQuery),
+		MemoryBytesQuotaPerQuery:        int64(opts.MemoryBytesQuotaPerQuery),
+		MaxMemoryBytes:                  int64(opts.MaxMemoryBytes),
 		QueueSize:                       opts.QueueSize,
 		ExecutorDependencies:            dependencyList,
 		FluxLogEnabled:                  opts.FluxLogEnabled,
